requirements.txt
Django>=4.2
gunicorn>=20.1.0
psycopg2-binary
django-redis>=5.2.0
celery>=5.2.7 
python-decouple>=3.7 
whitenoise>=6.2.0 

pillow==10.2.0
django-solo==2.2.0
djangorestframework==3.15.1
django-cors-headers==4.3.1
# coreapi==2.3
django-admin-sortable2==2.1
openpyxl==3.1.2
django-import-export==4.0.3
